all clear

+++ consumer/models.py
@@ -64,20 +64,21 @@ class Resume(models.Model):
     # 이미지 이름 바꾸기 코드 짜기
 
     def __str__(self):
-        return str(self.author)
+        return str(self.title)
 
     def save(self, *args, **kwargs):
         if self.pk:
             original_resume = Resume.objects.get(pk=self.pk)
-            if original_resume.mainimage.name != self.mainimage.name:
+            if original_resume.mainimage and original_resume.mainimage.name != self.mainimage.name:
                 storage.delete(original_resume.mainimage.name)
-            if original_resume.otherimages1.name !=self.otherimages1.name:
+            if original_resume.otherimages1 and original_resume.otherimages1.name != self.otherimages1.name:
                 storage.delete(original_resume.otherimages1.name)
-            if original_resume.otherimages2.name !=self.otherimages2.name:
+            if original_resume.otherimages2 and original_resume.otherimages2.name != self.otherimages2.name:
                 storage.delete(original_resume.otherimages2.name)
-            if original_resume.otherimages3.name !=self.otherimages3.name:
+            if original_resume.otherimages3 and original_resume.otherimages3.name != self.otherimages3.name:
                 storage.delete(original_resume.otherimages3.name)
-            if original_resume.otherimages4.name !=self.otherimages4.name:
+            if original_resume.otherimages4 and original_resume.otherimages4.name != self.otherimages4.name:
                 storage.delete(original_resume.otherimages4.name)
 
         super().save(*args, **kwargs)
+
